pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "weaselbench"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "jsonschema>=4.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-tmp-files>=0.0.2",
]

[project.scripts]
weaselbench = "weaselbench.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
weaselbench = ["web_assets/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
